src/data-processor/requirements.txt
opencv-python
mediapipe
google-cloud-aiplatform
pandas
numpy
tqdm
google-cloud-storage